Artur: To get more results the search is now listing all titles under £8. Tests have been changed to reflect that.
AmazonGoer class has a private variable that indicates the number of pages in a search - this should be retrieved automatically, as the number can vary (currently is around 350, amazon only allows up to 400 pages). This can be easily retrieved during the BluRayParser runtime. Change this number to specify number of pages to go through.

For some reason when the parser gets to page 2 it only processes first 5 titles and then stops, ignoring the rest of the titles on that page and the following pages.

listOfBluRays array should be declared outside of the method parseIntoBluRays, otherwise it gets overwritten each time a new page is being processed.